Fixes failing integration test. Section unit tests work again
