+++ internal/pkg/section/testUtil.go
@@ -63,6 +63,18 @@ func CheckZone(z1, z2 *Zone, t *testing.T) {
 	}
 }
 
+func CheckNotification(n1, n2 *Notification, t *testing.T) {
+	if n1.Type != n2.Type {
+		t.Error("Notification Type mismatch")
+	}
+	if n1.Token != n2.Token {
+		t.Error("Notification Token mismatch")
+	}
+	if n1.Data != n2.Data {
+		t.Error("Notification Data mismatch")
+	}
+}
+
 func CheckSignatures(s1, s2 []signature.Sig, t *testing.T) {
 	if len(s1) != len(s2) {
 		t.Error("Signature count mismatch")
